Created a service to save transactions from cms to pastel

+++ src/main/java/com/untucapital/usuite/utg/repository2/PostGlRepository.java
@@ -4,14 +4,19 @@ import com.untucapital.usuite.utg.entity.PostGl;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface PostGlRepository extends JpaRepository<PostGl, Long> {
+public interface PostGlRepository extends JpaRepository<PostGl, BigInteger> {
 //    List<PostGl> findByTxDate(Date TxDate);
-
+//
 //    List<PostGl> findByAccountLink(Integer AccountLink);
+
+
+
+
